workshops/veknow/requirements.txt
veadk-python
requests
pycryptodome
lark-oapi
